requirements.txt
streamlit
openai
psycopg2-binary
requests
python-dotenv
google-generativeai
